💡 Handled updated API response for logged out member
refs https://github.com/TryGhost/Team/issues/560
refs https://github.com/TryGhost/Ghost/commit/6b6134c8c844108ba114167c0d8807a2762e6c56

The endpoint `/members/api/member/` is used by Portal for fetching member details on site load to setup different flows. The response from this endpoint for logged out member has now changed from 401 Unauthorized to 204 No Content. This updates the API handling in Portal for updated response status.

This change also handles the visible error messages on console for Portal when it first loads and tries to establish member logged-in status, and instead handles it now behind the scenes. This also makes Portal functioning more transparent on the site.