Removed setting existing legacy prices to inactive
refs https://github.com/TryGhost/Members/commit/bb3b167d23d1e4a4cf326ef0fb458ffde3d0eb4a

The current implementation of setting legacy prices as inactive has a bug where it also sets the active price to inactive if only one of monthly/yearly price is edited while the other is not changed. This is a temporary reversal till we can narrow down the issue and fix to correctly set legacy plans to inactive.