🐛 Fixed missing complimentary subscription add button
refs https://github.com/TryGhost/Ghost/commit/84e4e754b214c9278dc2b865a2fb9ed4449c8600
refs https://github.com/TryGhost/Team/issues/758

The "Add complimentary" subscription button in members does not show up when members already have an existing zero amount subscription. But this was incorrectly not taking into account active subscriptions and was applying the rule to canceled subscriptions. Since the `comped` behaviour changed in 4.6 which caused member's existing comp subscription to be canceled, this bug did not allow the comped subscription to be added back.