Restricted Stripe Checkout to Members without products
refs https://github.com/TryGhost/Team/issues/858

Previously when complimentary access used a Stripe subscription under
the hood, we were able to restrict access to Stripe Checkout based on
whether or not a Member had an active Stripe subscription. This updates
the logic in members-api to instead check for whether or not the member
has 1-or-more products.