Fixed `created_at` filter not applied on members list query
refs https://github.com/TryGhost/Admin/commit/360db2526daed1ffe0d4b11caae05ed17a4232bd

Since the concat method does not change the existing array but instead returns a new array, the extra filters on the API query object were not being applied. The extra filter included filtering on `created_at` date for members which was not applied as the filter was not stored in the array correctly.